Fixed: crash when adding a texture to a new map Fixed: some bugs caused by refactoring Fixed: waterheight is set to 1 instead of -1 (stupid RO negative height thing)

+++ src/wateredit.cpp
@@ -20,7 +20,7 @@ int cProcessManagement::wateredit_process_events(SDL_Event &event)
 				cGraphics::world->water.type = max(0, cGraphics::world->water.type - 1);
 				break;
 			case SDLK_RIGHTBRACKET:
-				cGraphics::world->water.type = min(cGraphics::waterCount-1, cGraphics::world->water.type + 1);
+				cGraphics::world->water.type = min((int)cGraphics::waterCount-1, cGraphics::world->water.type + 1);
 				break;
 			case SDLK_PAGEDOWN:
 				cGraphics::worldContainer->undoStack->push(new cUndoChangeWater(cGraphics::world->water));
